Moving the IOC registration code into src and referencing the code from the transpiled `lib/` directory.
