+++ registerConfig.js
@@ -1,20 +1,1 @@
-const spurConfig = require('spur-config');
-
-module.exports = function (ioc, folderPath, configNameArg) {
-  try {
-    const configName = configNameArg || 'config';
-    const configLoaderName = '${configName}Loader';
-
-    const configLoader = spurConfig.load(folderPath);
-    const config = configLoader.getConfig();
-
-    const dependencies = {};
-    dependencies[configLoaderName] = configLoader;
-    dependencies[configName] = config;
-
-    ioc.registerDependencies(dependencies);
-  } catch (e) {
-    console.error('Error registering config', e);
-    throw e;
-  }
-};
+module.exports = require('./lib/iocRegistration');
